Move non-middleware out of middleware folder
